requirements.txt
flask~=1.1.2
requests~=2.24.0
Flask-SQLAlchemy
Flask-Restful
flask-restplus
flask-marshmallow
passlib~=1.7.4
PyJWT~=1.7.1
flask_httpauth
flask-login
flask-wtf
flask-mail
sqlalchemy
pymsteams~=0.1.13
urllib3~=1.25.9
gunicorn
Werkzeug==0.16.0
marshmallow-sqlalchemy
WTForms~=2.3.3
email-validator
itsdangerous~=1.1.0
Flask-Migrate==2.5.3